Changes around buggy preferences
There were a lot of overlooked codes - setting what must not be set but
forgetting to set what is needed etc. This is fixed now. Also OK/Cancel
is implemented. There's only small glitch - account list is not live -
but this could be easily handled.

Also UI is stil not handled properly because of problem with
nsISupportedArray used in JavaScript (probably and hopefully)