Rewritten Makefile to better server developers
Previous revisiou of Makefile was buggy - it didn't use xpidl
correctly and generated XPT files weren't complete. And addition it
does everything in one task and doesn't help developers which is now
better thanks to 'xpts' target defined in Makefile.

If developer changes IDL definition, everything she needs is to run
`make xpts` in and XPTs needed by the extension are updated properly.