Finally solved passing parameters right to XML-RPC client
Previous solutions was just a bluff on XPConnect but it failed in
XML-RPC when request was build. This solution works and includes fixes
on XML-RPC clien itself - it is possible that nsISupportsPrimitive
implemetation change over time. Anyway, this is the reason to
transition to nsIVariant and more systematic changes on the client.