Module registration and service initialization
There are lots of fixes to make everything compile after module
addition. Changes also tried to solve existence of Init and Shutdown
methods of DND TXT Service which are quite problematic because they're
obsolete in the Necko library. I chose the same way as was originally
made in Necko - intermediate interface nsPIDNSTXTService. It seems
that it shoudld work quite seemlesly but we can find out only after
some application.

URL Helper was removed because it wasn't really needed. Actually it
was responsible for hostname validation in the original code and it we
could modify to validate TXT records - but this doesn't make sense
(they're as a free text there).

EEE-specific TXT records are not recognized in our DNS TXT module so
it shouldn't bother us to loose URL Helper.