Method call replaced with variable
Because Mozilla doesn't really obey basic OOP principles (considered
stupid by some) and we copied a lot of code from Mozilla,
GetNextRecord move internal and returns value. This means that we
can't call GetNextRecord multiple times and expect the same result and
therefore we must introduce temp variable.

Temp variable was there but in this particular case, I forgot to
replace method call with it.