Improved DNS API based Ruby resolv
Resolver itself is still very similar to C code. Problem is that
js-ctypes apparently doesn't provide any way of pointer arythmetics
and because of that current implementation needs to be rewritten.

New version should be based on char array manipulations which is far
from being C-like and probably quite inefficient which actually
shouldn't matter.