Client uses lexical scope instead of 'this'
This main change is accompanied by few others.  The whole file is
compliant with Google Closure Linter which is obviously a direction we
want to take as it greatly helps new commers to write valid code.
Services code module was introduced to simplify Mozilla service calls.
Old parameter serialization for queue was removed because it's no
longer necessary.  Timer is observer through ad-hoc XPCOM object which
is another change we're going to introduce all around the extension.