Opening eee attachments in browser
When user doubleclicks on attachment with eee uri it is opened by web browser.
Translating eee uri to http uri isn't completed because dns implementation isn't completed and it is
needed in translation to get server hostname. Meanwhile as server hostname is taken server name as
is in eee uri.