Synchronization module reworked
Queue was renamed to Method.  Because what it really does is making
method calls synchronized.

There is also new Queue.  Which synchroniously executes pushed
functions.  When one fuction is done, it calls the next one, but
doesn't have to know what function it is.  This is necessary for
nonblocking authentization which is major performance problem now.  It
also opens the doors for greater simplification of basic protocol
usage scenarios.