Rewritten synchronization
There was a problem with Synchronization.Queue sharing because,
obviously, some scenarios were triggered at nearly the same time and
locks haven't been in effect yet.  So it is now created every time
scenario is called.

This had created an opportunity and a problem.  The opportunity was to
move inner fuction from createScenario to right to Request.Client
constructor.  The problem was that we still needed to pass control to
the next function in queue.  I've solved both, but the solution to the
problem is not good (I have some other ideas to try).