Authentication delegate rewritten
It was mainly moved back a bit, because nsIAuthPromp#promptPassword
we're using at the moment doesn't store password for session.  So must
still handle quite a lot of heavy lifting around password (i.e., login
info) management.

The only thing missing is queue validation for cases when user clicked
cancel and doesn't want to be bothered by password prompts for the
rest of the session.

It also seems like there can still be multiple password prompts at the
same time.  And that authentication can be skipped.  Altough I'm not
able to reproduce the conditions.