Queue validation refactored
It was move to a separate delegate.  This allowed cal3eRequest.Client
to be simplified even further.

Regarding validation itself, it is not important what user error code
are we looking for.  What matter is whether there's any valid user
error which is simpler and logical.  So the algorithm was changed and
refactored too - no temp value, simple function - all the good stuff.