Fixed issue with blocking parallel execution
This has probably negative performance issue and other strategy should
be found in the future.  It basically ensures that only one scenario
is executed at the time.  This for example means that multiple
ESClient.queryObjects calls are called one after another resulting in
slow data loads (especially during initialization).

However, the fix was extremely simple and it was probably forgotten
line of code during initial implementation.