Cache added to DNS module
It caches EEE TXT records returned from Resolv module.  To keep cache
visible with minimal impact on code it was necessary to refactor the
module.

Default record was moved from before-return place to records parsing
and filtering.  This also enables us to cache this default for one for
one day.

Records are cached with TTL of the smallest TTL in the record set.