No calendars can be subscribed fix
Subscription dialog wasn't able to handle this situation because it
didn't track whether structures are loaded.  It also didn't track
whether owner which were loaded are actually the ones we care about
which caused a problem because owners structure was the one to
determine whether everything is fine.

There's other possibility to skip loading users when query is empty
(meaning all shared calendars are subscribed or there are no shared
calendars).  It wasn't chosen at the moment because it doesn't cause
any significant performance issue but it would add a complexity to
current execution flow.