Logging improvements added to XML-RPC
It was changed how response is logged so that thrown expception
doesn't prevent debugging message with the whole response to be
logged.

New logger was also added to provide parser logging.  It is extremely
verbose even on info level, but it's a parser and XML is verbose on
its own.