SD tries and defaults implemented
This is basically a callback onion (yes, nobody likes them) where
original callback is wrapped in by additional callbacks.  It should be
refactor to some queue because this abusing functional aspects of
JavaScript and it is ugly.  The basic principle should stay, create
something which checks condition and passed the result to next similar
something and so on until the next something is the original callback.

This way multiple execution can be implemented, default result
intercaption, cache setting etc.