SD default reworked
This resulted in changes in DNS parsing which are not relying on
RegExp anymore but properly follows updated EEE specification.

There's a new constructor which holds data about discovered service
like host, port and TTL.  It handles defaults itself so there's no
need for extra on this front.  This enabled simpler and more
straightforward DNS parsing and removal of extra queue in the main
service discovery object.