Implements well-known URI support
The implementation uses XHR and takes Cache-Control or Expires
response headers as source of TTL value.  Custom channel callback are
also used to support HTTPS properly in case redirects are in place.
Generally, mechanism is very similar to the on used in XML-RPC module.

Raw data to Service constructed object are not DRY at the moment.