Different approach to async DNS resolv
The WorkerResolverClient and WorkerResolverServer were introduced.
They both try to mitigate difficulty of using worked in extensions.
ChromeResolvClient is somewhat hardcorded to resolv.jsm but
ChromeResolvServer is pretty general.  The worker client "encodes"
messages which will be turned to method calls by worker server.

The old Resolv.DNS was forces to be asychronous and uses futures as
its result which radically simplifies its use.

The resolvers themselves stopped to handle worker and don't care about
which made them less coupled to stuff they don't need.  This is
possible thanks to WorkerResolverServer.