On behalf events in subscribed calendars
When meetings are created in subscribed calendar they'll use
calendar's owner as organizer, but subscribee is marked as acting on
behalf of the owner (using a SENT-BY parameter).

The implementation is not using ACL manager, it is modifying event
passed to addItem, adoptItem or modifyItem.  So, it is pretty dumb,
but it is backwards compatible down to Postbox and from the
perspective of synchronizer it is easier to achieve using just
organizerId and organizerCN properties (i.e., the so called old way).