Fixed organizer among attendee before sync
This is fix of workarounf if this bug https://bugzilla.mozilla.org/show_bug.cgi?id=459683

Before event is sent to server organizer is removed from attendee list. This modified
event was also returned back to lightning. So when you opened event before sync with server
it listed organizer among attendees but after sync it did not.